[refactor]: Button 을 common 으로 이동 (#23)
