+++ app/src/main/java/com/pickpoint/pickpoint/ui/common/component/ResetConfirmButton.kt
@@ -1,4 +1,4 @@
-package com.pickpoint.pickpoint.ui.randompicker.component
+package com.pickpoint.pickpoint.ui.common.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
@@ -10,8 +10,6 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -82,44 +80,9 @@ fun ResetConfirmButton(reset: () -> Unit, confirm: () -> Unit) {
 
 }
 
-@Composable
-fun RetryButton(retry:() -> Unit) {
-
-    Button(
-        onClick = { retry() },
-        colors = ButtonDefaults.buttonColors(
-            containerColor = Color(0xFF333333),
-            contentColor = Color(0xFFFFFFFF),
-        ),
-        shape = RoundedCornerShape(100.dp),
-        modifier = Modifier.width(300.dp)
-            .height(48.dp)
-            .shadow(4.dp, RoundedCornerShape(100.dp))
-    ) {
-        Icon(
-            painter = painterResource(id = R.drawable.ic_retry),
-            contentDescription = "Cancel"
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-        Text(
-            "Retry",
-            style = TextStyle(
-                fontSize = 20.sp
-            )
-        )
-    }
-
-}
-
 
 @Preview
 @Composable
 fun ResetConfirmButtonPreview() {
     ResetConfirmButton(reset = { /* Handle reset click */ }, confirm = { /* Handle confirm click */ })
-}
-
-@Preview
-@Composable
-fun RetryButtonPreview() {
-    RetryButton(retry = { /* Handle retry click */ })
 }
